My Exercises page now displaying personalised data based on dynamic routing
